@earowang 2019-08-14
Can you also print out the fable output? just want to make sure there's no `*` in the output.