@mitchelloharawild 2019-08-14
``` r
library(fable)
#> Loading required package: fabletools
library(tsibble)
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following object is masked from 'package:tsibble':
#> 
#>     id
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

fits <- model(tourism, SNAIVE(Trips))
forecast(fits, h = 1)
#> # A fable: 304 x 7 [1Q]
#> # Key:     Region, State, Purpose, .model [304]
#>    Region     State       Purpose  .model     Quarter   Trips .distribution
#>    <chr>      <chr>       <chr>    <chr>        <qtr>   <dbl> <dist>       
#>  1 Adelaide   South Aust… Business SNAIVE(T…  2018 Q1 129.    N(129, 2018) 
#>  2 Adelaide   South Aust… Holiday  SNAIVE(T…  2018 Q1 216.    N(216, 794)  
#>  3 Adelaide   South Aust… Other    SNAIVE(T…  2018 Q1  93.8   N(94, 344)   
#>  4 Adelaide   South Aust… Visiting SNAIVE(T…  2018 Q1 270.    N(270, 1309) 
#>  5 Adelaide … South Aust… Business SNAIVE(T…  2018 Q1   0.979 N(0.98, 40)  
#>  6 Adelaide … South Aust… Holiday  SNAIVE(T…  2018 Q1  12.9   N(13, 64)    
#>  7 Adelaide … South Aust… Other    SNAIVE(T…  2018 Q1   0.684 N(0.68, 4.8) 
#>  8 Adelaide … South Aust… Visiting SNAIVE(T…  2018 Q1  22.4   N(22, 185)   
#>  9 Alice Spr… Northern T… Business SNAIVE(T…  2018 Q1  18.6   N(19, 61)    
#> 10 Alice Spr… Northern T… Holiday  SNAIVE(T…  2018 Q1  21.1   N(21, 148)   
#> # … with 294 more rows
```

<sup>Created on 2019-08-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>